reports-service/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# GraphQL
strawberry-graphql[fastapi]==0.215.1

# Database
sqlalchemy>=1.4.0,<2.0.0
pymysql==1.1.0
cryptography==41.0.7

# Excel Export
openpyxl==3.1.2

# SOAP Client
zeep==4.2.1

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0

# Utilities
pydantic>=1.10.0,<2.0.0
pydantic[dotenv]>=1.10.0,<2.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2